big-f format cleaned up backend
