+++ tests/functional_tests.py
@@ -1,8 +1,11 @@
+import logging
 import os
 from typing import Union
 import unittest
 
-from pyBIG import Archive, LargeArchive
+from pyBIG import InMemoryArchive, InDiskArchive
+
+logging.basicConfig(level=logging.DEBUG)
 
 
 TEST_FILE = "read_me_for_test.txt"
@@ -13,14 +16,14 @@ TEST_ARCHIVE = "tests/test_data/empty_archive.big"
 
 class BaseTestCases:
     class BaseTest(unittest.TestCase):
-        archive: Union[Archive, LargeArchive]
+        archive: Union[InMemoryArchive, InDiskArchive]
 
         def tearDown(self):
             for file in [
                 f"tests/test_data/output/{TEST_FILE}",
                 "tests/test_data/output/test.big",
                 TEST_ARCHIVE,
-                "tests/test_data/test_big_type.big"
+                "tests/test_data/test_big_type.big",
             ]:
                 try:
                     os.remove(file)
@@ -66,7 +69,7 @@ class BaseTestCases:
             self.archive.extract("tests/test_data/output")
             self.assertTrue(os.path.exists(f"tests/test_data/output/{TEST_FILE}"))
 
-            new_archive = Archive.from_directory("tests/test_data/output")
+            new_archive = InMemoryArchive.from_directory("tests/test_data/output")
             self.assertIn(TEST_FILE, new_archive.entries)
 
             new_archive.save("tests/test_data/output/test.big")
@@ -79,10 +82,10 @@ class BaseTestCases:
 class TestArchive(BaseTestCases.BaseTest):
     def setUp(self):
         with open("tests/test_data/test_big.big", "rb") as f:
-            self.archive = Archive(f.read())
+            self.archive = InMemoryArchive(f.read())
 
     def test_empty_archive(self):
-        archive = Archive.empty()
+        archive = InMemoryArchive.empty()
         archive.add_file(TEST_FILE, TEST_CONTENT.encode(TEST_ENCODING))
         archive.repack()
 
@@ -92,12 +95,12 @@ class TestArchive(BaseTestCases.BaseTest):
         path = "tests/test_data/test_big_type.big"
 
         for header in ["BIG4", "BIGF"]:
-            archive = Archive.empty(header=header)
+            archive = InMemoryArchive.empty(header=header)
             archive.add_file(TEST_FILE, TEST_CONTENT.encode(TEST_ENCODING))
             archive.save(path)
 
             with open(path, "rb") as f:
-                archive = Archive(f.read())
+                archive = InMemoryArchive(f.read())
 
             self.assertEqual(archive.header, header)
             os.remove("tests/test_data/test_big_type.big")
@@ -105,7 +108,7 @@ class TestArchive(BaseTestCases.BaseTest):
 
 class TestLargeArchive(BaseTestCases.BaseTest):
     def setUp(self):
-        self.archive = LargeArchive("tests/test_data/test_big.big")
+        self.archive = InDiskArchive("tests/test_data/test_big.big")
 
         if self.archive.file_exists(TEST_FILE):
             self.archive.remove_file(TEST_FILE)
@@ -117,14 +120,14 @@ class TestLargeArchive(BaseTestCases.BaseTest):
         self.archive.repack()
 
     def test_empty_archive(self):
-        archive = LargeArchive.empty(TEST_ARCHIVE)
+        archive = InDiskArchive.empty(TEST_ARCHIVE)
         archive.add_file(TEST_FILE, TEST_CONTENT.encode(TEST_ENCODING))
         archive.repack()
 
         self.assertIn(TEST_FILE, archive.entries)
 
     def test_archive_memory_size(self):
-        archive = LargeArchive.empty(TEST_ARCHIVE)
+        archive = InDiskArchive.empty(TEST_ARCHIVE)
         file_bytes = TEST_CONTENT.encode(TEST_ENCODING)
         archive.add_file(TEST_FILE, file_bytes)
 
@@ -136,10 +139,10 @@ class TestLargeArchive(BaseTestCases.BaseTest):
         path = "tests/test_data/test_big_type.big"
 
         for header in ["BIG4", "BIGF"]:
-            archive = LargeArchive.empty(path, header=header)
+            archive = InDiskArchive.empty(path, header=header)
             archive.add_file(TEST_FILE, TEST_CONTENT.encode(TEST_ENCODING))
             archive.save()
-            archive = LargeArchive(path)
+            archive = InDiskArchive(path)
             self.assertEqual(archive.header, header)
             os.remove("tests/test_data/test_big_type.big")
