+++ tests/memory_tests.py
@@ -1,18 +1,18 @@
 import os
 import unittest
 
-from pyBIG import Archive, LargeArchive
+from pyBIG import InMemoryArchive, InDiskArchive
 from pyBIG.utils import getsize
 
 
 class MemoryTests(unittest.TestCase):
     def test_1(self):
-        archive = Archive.empty()
+        archive = InMemoryArchive.empty()
         archive.add_file("test_file.txt", b"Clement1" * 1250000)
 
         in_memory_size = getsize(archive)
 
-        archive = LargeArchive.empty("big_big.big")
+        archive = InDiskArchive.empty("big_big.big")
         archive.add_file("test_file.txt", b"Clement1" * 1250000)
         archive.save()
 
@@ -23,13 +23,13 @@ class MemoryTests(unittest.TestCase):
         assert in_memory_size > on_disk_size
 
     def test_2(self):
-        archive = LargeArchive.empty("big_big.big")
+        archive = InDiskArchive.empty("big_big.big")
         archive.add_file("test_file.txt", b"Clement1" * 1250000)
         archive.save()
 
         post_save_size = getsize(archive)
 
-        archive = LargeArchive("big_big.big")
+        archive = InDiskArchive("big_big.big")
 
         loaded_size = getsize(archive)
